+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - [797](https://github.com/vegaprotocol/core-test-coverage/issues/797) - Add coverage for `0011-MARA-030`
 - [798](https://github.com/vegaprotocol/core-test-coverage/issues/798) - Add coverage for `0011-MARA-031`
 - [799](https://github.com/vegaprotocol/core-test-coverage/issues/799) - Add coverage for `0011-MARA-032`
+- [10392](https://github.com/vegaprotocol/vega/issues/10392) - Integrate mark price with price monitoring.
 
 ### 🐛 Fixes